header component created, navbar adjusted, partially styled
